countdec removed and count added in OeCalculator.java & RcCalculator.java

+++ pojo-ops/src/main/java/com/stanzaliving/core/opscalculator/dto/CategoryDto.java
@@ -23,7 +23,7 @@ import lombok.experimental.SuperBuilder;
 @AllArgsConstructor
 public abstract class CategoryDto {
 
-	private int count;
+	private double count;
 	private int rate;
 	private int utilization;
 	private int margin;
